requirements.txt
ipykernel
numpy
pandas
scikit-learn
imbalanced-learn
bentoml
pydantic
requests
pyjwt
pathlib 
joblib
flake8
fastapi
datetime